requirements.txt
matplotlib>=3.9.0,<3.10
mysql-connector-python>=9.0.0
pytest>=8.0.0
